removing cyclic dependency
